Update and move logic to Repository
Implement the PaymentQueryService functionality #153
